Refactor daily tip to routeLodaer
